requirements_web.txt
# Web框架
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# Grounding DINO (需要从源码安装)
# 请参考: https://github.com/IDEA-Research/GroundingDINO

# VLM客户端
openai>=1.3.0
httpx>=0.25.0

# 数据处理
pandas>=2.0.0
pyyaml>=6.0

# 工具库
tqdm>=4.66.0
python-multipart>=0.0.6
aiofiles>=23.2.0
requests>=2.31.0

# Label Studio 集成
label-studio-sdk>=0.0.24

# 文件处理
zipfile36>=0.1.3

